+++ packages/framework/src/Helpers/HydeHelperFacade.php
@@ -2,6 +2,8 @@
 
 namespace Hyde\Framework\Helpers;
 
+use Illuminate\Support\Str;
+
 /**
  * Provides convenient access to Hyde helpers, through the main Hyde facade.
  *
@@ -17,5 +19,13 @@ trait HydeHelperFacade
     public static function hasFeature(string $feature): bool
     {
         return Features::enabled($feature);
+    }
+
+    /**
+     * @since 0.44.0-beta (renamed from titleFromSlug)
+     */
+    public static function makeTitle(string $slug): string
+    {
+        return Str::title(str_replace('-', ' ', ($slug)));
     }
 }